requirements.txt
typer[all]>=0.9.0
rich>=13.0.0

PyYAML>=6.0

requests>=2.31.0

psutil>=5.9.0

python-dateutil>=2.8.2
